Split TestDowncall into multiple test classes + relax rejection of call types
